# application/api-spec/src/main/resources/openapi/rental-api.yaml
openapi: 3.0.0
info:
  title: Property Rental API
  version: 1.0.0
  description: API for managing property rental listings
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
  /api/listings:
    get:
      summary: Search active listings with optional filters
      operationId: searchListings
      tags:
        - Listing
      parameters:
        - name: city
          in: query
          required: false
          schema:
            type: string
        - name: listingType
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ListingType'
        - name: priceMin
          in: query
          required: false
          schema:
            type: number
            format: double
        - name: priceMax
          in: query
          required: false
          schema:
            type: number
            format: double
        - name: propertyType
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/PropertyType'
        - name: roomCount
          in: query
          required: false
          schema:
            type: integer
        - name: areaMin
          in: query
          required: false
          schema:
            type: number
            format: double
        - name: areaMax
          in: query
          required: false
          schema:
            type: number
            format: double
        - name: furnished
          in: query
          required: false
          schema:
            type: boolean
        - name: parkingAvailable
          in: query
          required: false
          schema:
            type: boolean
        - name: balcony
          in: query
          required: false
          schema:
            type: boolean
        - name: petsAllowed
          in: query
          required: false
          schema:
            type: boolean
        - name: energyClass
          in: query
          required: false
          schema:
            type: string
            enum: [A, B, C, D]
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
            enum: [price_asc, price_desc, newest, area_asc, area_desc]
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Listings matching the filters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingSearchResponse'
        '401':
          description: Unauthorized - Missing or invalid JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a new property listing
      operationId: createListing
      tags:
        - Listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateListingRequest'
      responses:
        '201':
          description: Listing successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingResponse'
        '400':
          description: Bad Request - Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Missing or invalid JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - User is not an OWNER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/listings/featured:
    get:
      summary: Get featured listings, optionally filtered (public endpoint for home page)
      operationId: getFeaturedListings
      tags:
        - Listing
      security: []
      parameters:
        - name: city
          in: query
          required: false
          schema:
            type: string
        - name: listingType
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ListingType'
        - name: propertyType
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/PropertyType'
      responses:
        '200':
          description: Featured listings (limited list, can be empty)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ListingSummary'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/listings/my:
    get:
      summary: List listings owned by the current authenticated user
      operationId: getMyListings
      tags:
        - Listing
      responses:
        '200':
          description: Listings owned by current user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ListingResponse'
        '401':
          description: Unauthorized - Missing or invalid JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/user:
    get:
      summary: Get current authenticated user's profile
      operationId: getCurrentUser
      tags:
        - User
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: Update current user's profile
      operationId: updateUserProfile
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserProfileRequest'
      responses:
        '200':
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/user/avatar:
    post:
      summary: Upload or replace the current user's avatar photo
      operationId: uploadUserAvatar
      tags:
        - User
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file to upload (max 10 MB)
      responses:
        '200':
          description: Avatar updated, returns updated user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/listings/{listingId}/photos:
    get:
      summary: List photos for a listing
      operationId: getListingPhotos
      tags:
        - Listing
      security: []
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of photos for the listing
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PhotoResponse'
        '404':
          description: Listing not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Upload a photo for a listing (owner only)
      operationId: uploadListingPhoto
      tags:
        - Listing
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file to upload (max 10 MB)
                altText:
                  type: string
                  description: Optional alt text for the image
      responses:
        '201':
          description: Photo uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhotoResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - not the listing owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Listing not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/listings/{id}:
    get:
      summary: Get listing by ID
      operationId: getListingById
      tags:
        - Listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Listing found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingResponse'
        '401':
          description: Unauthorized - Missing or invalid JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Listing not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    CreateListingRequest:
      type: object
      required:
        - title
        - description
        - listingType
        - address
        - price
        - features
      properties:
        title:
          type: string
          description: Title of the listing
          example: "Modern 2-bedroom apartment in city center"
        description:
          type: string
          description: Detailed description of the property
          example: "Beautiful apartment with great view"
        listingType:
          $ref: '#/components/schemas/ListingType'
        address:
          $ref: '#/components/schemas/AddressRequest'
        price:
          $ref: '#/components/schemas/PriceRequest'
        features:
          $ref: '#/components/schemas/PropertyFeaturesRequest'
...
    ListingSummary:
      type: object
      required: [ id, title, listingType, city, price ]
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        listingType:
          $ref: '#/components/schemas/ListingType'
        city:
          type: string
        price:
          $ref: '#/components/schemas/PriceResponse'
        coverPhoto:
          $ref: '#/components/schemas/PhotoResponse'
    ListingResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: "Modern 2-bedroom apartment in city center"
        description:
          type: string
          example: "Beautiful apartment with great view"
        listingType:
          $ref: '#/components/schemas/ListingType'
        status:
          $ref: '#/components/schemas/ListingStatus'
        createdAt:
          type: string
          format: date-time
          example: "2025-04-03T10:00:00Z"
        address:
          $ref: '#/components/schemas/AddressResponse'
        price:
          $ref: '#/components/schemas/PriceResponse'
        features:
          $ref: '#/components/schemas/PropertyFeaturesResponse'
        ownerId:
          type: integer
          format: int64
          example: 123
        owner:
          $ref: '#/components/schemas/User'
        photos:
          type: array
          items:
            $ref: '#/components/schemas/PhotoResponse'
    PhotoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        altText:
          type: string
        contentType:
          type: string
          example: image/jpeg
        originalFilename:
          type: string
          example: living-room.jpg
        position:
          type: integer
        contentUrl:
          type: string
          example: /api/listings/1/photos/10/content
    AddressRequest:
      type: object
      required:
        - street
        - city
        - postalCode
        - country
      properties:
        street:
          type: string
          example: "Main Street 123"
        city:
          type: string
          example: "Bratislava"
        postalCode:
          type: string
          example: "81101"
        country:
          type: string
          example: "Slovakia"
        district:
          type: string
          example: "Old Town"
        region:
          type: string
          example: "Bratislava Region"
        lat:
          type: number
          format: double
          example: 48.1486
        lng:
          type: number
          format: double
          example: 17.1077
    AddressResponse:
      type: object
      properties:
        street:
          type: string
          example: "Main Street 123"
        city:
          type: string
          example: "Bratislava"
        postalCode:
          type: string
          example: "81101"
        country:
          type: string
          example: "Slovakia"
        district:
          type: string
          example: "Old Town"
        region:
          type: string
          example: "Bratislava Region"
        lat:
          type: number
          format: double
          example: 48.1486
        lng:
          type: number
          format: double
          example: 17.1077
    PriceRequest:
      type: object
      required:
        - amount
        - currency
      properties:
        amount:
          type: number
          format: double
          example: 850.00
        currency:
          type: string
          example: "EUR"
    PriceResponse:
      type: object
      properties:
        amount:
          type: number
          format: double
          example: 850.00
        currency:
          type: string
          example: "EUR"
    PropertyFeaturesRequest:
      type: object
      required:
        - propertyType
      properties:
        propertyType:
          $ref: '#/components/schemas/PropertyType'
        area:
          type: number
          format: double
          description: Area in square meters
          example: 65.5
        roomCount:
          type: integer
          example: 2
        floor:
          type: integer
          example: 3
        furnished:
          type: boolean
          example: true
        parkingAvailable:
          type: boolean
          example: false
        balcony:
          type: boolean
          example: true
        elevator:
          type: boolean
          example: false
        petsAllowed:
          type: boolean
          example: true
        energyClass:
          type: string
          example: "A"
        yearBuilt:
          type: integer
          example: 2015
    PropertyFeaturesResponse:
      type: object
      properties:
        propertyType:
          $ref: '#/components/schemas/PropertyType'
        area:
          type: number
          format: double
          example: 65.5
        roomCount:
          type: integer
          example: 2
        floor:
          type: integer
          example: 3
        furnished:
          type: boolean
          example: true
        parkingAvailable:
          type: boolean
          example: false
        balcony:
          type: boolean
          example: true
        elevator:
          type: boolean
          example: false
        petsAllowed:
          type: boolean
          example: true
        energyClass:
          type: string
          example: "A"
        yearBuilt:
          type: integer
          example: 2015
    ListingType:
      type: string
      enum:
        - RENT
        - SALE
      example: RENT
    ListingStatus:
      type: string
      enum:
        - ACTIVE
        - INACTIVE
      example: INACTIVE
    PropertyType:
      type: string
      enum:
        - APARTMENT
        - HOUSE
        - ROOM
      example: APARTMENT
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "John"
        surname:
          type: string
          example: "Doe"
        email:
          type: string
          example: "owner@example.com"
        phone:
          type: string
          example: "+421 900 000 000"
        bio:
          type: string
          example: "Looking for a flat in Bratislava."
        role:
          $ref: '#/components/schemas/UserRole'
        avatarUrl:
          type: string
          description: URL to the user's avatar image
          example: /api/photos/42/content
    UpdateUserProfileRequest:
      type: object
      required: [name, surname, email]
      properties:
        name:
          type: string
          example: "John"
        surname:
          type: string
          example: "Doe"
        email:
          type: string
          example: "john@example.com"
        phone:
          type: string
          example: "+421 900 000 000"
        bio:
          type: string
          example: "Looking for a flat in Bratislava."
    UserRole:
      type: string
      enum:
        - OWNER
        - USER
      example: OWNER
    ListingSearchResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ListingResponse'
        pagination:
          $ref: '#/components/schemas/PaginationResponse'
    PaginationResponse:
      type: object
      properties:
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 10
        totalElements:
          type: integer
          format: int64
          example: 42
        totalPages:
          type: integer
          example: 5
    ErrorResponse:
      type: object
      properties:
        type:
          type: string
          description: Type of error
          example: "VALIDATION"
        message:
          type: string
          description: Error message
          example: "Listing title is required."
        field:
          type: string
          description: Field name for validation errors (optional)
          example: "title"
        timestamp:
          type: string
          format: date-time
          example: "2025-04-03T10:00:00Z"